refactor: move tracking fn to utils, prevent tracking system bugs
