+++ src/pages/signup-affiliates/index.tsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useState } from 'react'
 import styled from 'styled-components'
-import trackEvent from './utils/_tracking'
-import { Submit } from './utils/_utils'
+import { Submit, trackEvent } from './utils/_utils'
 import { AffiliateAccountTypes, SignUpStatusTypes } from './_types'
 import { AffiliateSignupForm, AffiliateSignupStatus, Wizard } from './_lazy-loading'
 import { isBrowser } from 'common/utility'
@@ -93,6 +92,7 @@ const AffiliateSignup = () => {
         const handleBeforeUnload = (event) => {
             event.preventDefault()
             trackEvent({ action: 'close' })
+            trackEvent({ action: 'close_wizard' })
         }
         window.addEventListener('beforeunload', handleBeforeUnload)
         return () => window.removeEventListener('beforeunload', handleBeforeUnload)
@@ -113,8 +113,15 @@ const AffiliateSignup = () => {
 
     useEffect(() => {
         const partner_signup_error_message = affiliate_api_error?.error.message
-
-        if (partner_signup_error_message == 'Username not available') {
+        if (affiliate_api_data) {
+            trackEvent({
+                action: 'success_popup_opened',
+                user_choice:
+                    JSON.stringify(affiliate_api_error?.echo_req) ||
+                    'success, but without echo_req',
+            })
+            setSignupStatus('success')
+        } else if (partner_signup_error_message == 'Username not available') {
             trackEvent({ action: 'partners_signup_error', partner_signup_error_message })
             setSignupStatus(partner_signup_error_message)
         } else if (
@@ -126,15 +133,7 @@ const AffiliateSignup = () => {
             setSignupStatus('Your website is not a valid entry')
         } else if (partner_signup_error_message)
             trackEvent({ action: 'other_error', partner_signup_error_message })
-
-        if (affiliate_api_data) {
-            trackEvent({
-                action: 'success_popup_opened',
-                user_choice: JSON.stringify(affiliate_api_error?.echo_req) || 'default or null',
-            })
-            setSignupStatus('success')
-        }
-    }, [affiliate_api_data, affiliate_api_error, affiliateSend])
+    }, [affiliate_api_data, affiliate_api_error])
 
     useEffect(() => {
         setAffiliateAccount({
